Give a singular prompt and store input as an array
One prompt will appear on the page for the user to input an expression
with the first and last value being stored as numbers to pass to the
other functions. Output will be in the console. This is to test that
saving the input as an array will be a reliable method.