Create buttons and display for calculator in HTML
A basic skeleton with no formatting, styling, or functionality was
created to give a layout with which to work. Classes and ids were
assigned to each button for future adjustments. Dummy numbers were given
to the display to make sure it was output to the correct area on the
screen.