Add classes and ids in HTML and check for operator
New classes and ids were added to operators to allow easier access to
them in the script file.

A conditional was added to getNumber to check if an operator has been
selected yet. If not, then we know that we are taking in the left
operand. Otherwise, the right operand is being entered.