Add operator array, add elements to operator class
The splitArray function now splits the given array into two arrays. One
containing all the numbers entered by the user and the other with the
operators entered by the user.

Added the class .operator to more elements in HTML to allow them to be
entered into the numberOperator array for future operations to be
performed on.