Handle negative numbers, decimals and input length
Negative numbers can now be entered with the sign changing only once. If
you wish to convert the number back to a positive, you must clear and
start over. These negative numbers will be treated as any other number
as they are entered in the appropriate operations.

Decimals can now be entered and treated as numbers. The program will
round them to 4 decimal places to avoid uglier decimals appearing to
the user.

A maximum length of 20 characters may be entered into the calculator.
This is to prevent overloading the calculator.