Add background image and clean up calculations
The operate function was edited to allow more accurate calculations
after they have been reordered to fit the order of operations.

The number of characters entered by the user is now decremented when
they press the backspace. This allows the user to add characters until
the limit without being penalized for an erroneous entry.