Store user input as object for left operand
A new approach is being taken to solve the issue with order of
operations not working in some cases when the numbers and operators
were input in specific order. Numbers, operators and the result will
be stored in an object and only one operation on a set of numbers
will be allowed at at time.

To start, the left operand is stored in an object until the user enters
more than 9 characters and displays it to the UI. For now, only click
events are functioning. Once those are fully functional, the keyboard
will be added once more.